refactor(components): Relocate shipping controls
