+++ js/src/components/app-searchable-select-control/index.js
@@ -4,6 +4,9 @@
 import SearchableSelectControl from '../searchable-select-control';
 import './index.scss';
 
+/**
+ * Wrapper around SearchableSelectControl to apply consistent styling across the app.
+ */
 const AppSearchableSelectControl = ( props ) => {
 	return (
 		<SearchableSelectControl
